requirements.txt
python-telegram-bot
googletrans==4.0.0-rc1
beautifulsoup4
requests